Version 1.0.0
Delivered with a test page, the project only permit you to see how the
alerts appear and dissapear. The .js have to be configured. Actually, 2
"mods" are available : automatic close mod (that automatically close
alerts) and the manual ones. To change between mods, just set timeAppear
to 0 or 1-infinite. Jquery-2.1.4.min.js delivered with the project.