Version 1.2.0 - STABLE
And here it goes ! JQuery Online Alert is now totally working and has
been tested on multiple websites ! There is some things that have to be
fix or some features that will be available in the futur :
- Actually, JOA is set up to take every 3.5 seconds all alerts that have
been posted in the 5 previous seconds. But if a call is made, for
example, at 19:45:01, the system will take all alerts between 19:45:00
and 19:45:01 (the range of 5 seconds has been reduced to 1 second). It
will be fix next time.
- Adding a fadeIn() effect and droping down event on alert appearing.
- More translations.
- A system to take care of timezones between the client and the server.

A "How to Install" will appear on the next days.